Create (3) SMILES to morgen fingerprint
The first column of the table is ‘Compound’ and the second column is ‘SMILES’ for import xlsx. file.
The output Morgan fingerprint is a 1024-bit molecular Morgan fingerprint feature, the first column is the name of the molecule ‘Compound’, the second to the last column is the ‘1’ or ‘0’ of the 1024-bit Morgan fingerprint. ‘. 
Note: Please give them labels (1 or 0) in the last column.