Create (6) Model prediction
The training set imported by the machine learning prediction code needs to be exactly the same as when it was trained, and the format needs to be consistent when it is predicted, which means that the first column of the predicted file should be the molecule name, and the second-to-last column should be the 1024-bit Morgan fingerprint feature. The final output will be the predicted probabilities of the four models.